#155 rename: 장소 오류 제보 place 도메인으로 옮기기
