+++ src/main/java/org/sopt/solply_server/domain/place/dto/response/PlaceReportCreateResponse.java
