+++ src/main/java/org/sopt/solply_server/domain/place/entity/PlaceReport.java
