+++ src/main/java/org/sopt/solply_server/domain/place/service/PlaceReportService.java
@@ -3,20 +3,15 @@ package org.sopt.solply_server.domain.report.service;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.sopt.solply_server.domain.place.entity.Place;
-import org.sopt.solply_server.domain.report.dto.request.PlaceReportCreateRequest;
-import org.sopt.solply_server.domain.report.dto.response.PlaceReportCreateResponse;
-import org.sopt.solply_server.domain.report.entity.PlaceReport;
+import org.sopt.solply_server.domain.place.dto.request.PlaceReportCreateRequest;
+import org.sopt.solply_server.domain.place.dto.response.PlaceReportCreateResponse;
+import org.sopt.solply_server.domain.place.entity.PlaceReport;
 import org.sopt.solply_server.domain.report.repository.PlaceReportRepository;
 import org.sopt.solply_server.domain.report.util.PlaceReportValidator;
 import org.sopt.solply_server.domain.user.entity.User;
-import org.sopt.solply_server.global.exception.BusinessException;
-import org.sopt.solply_server.global.exception.ErrorCode;
 import org.sopt.solply_server.global.util.EntityLoader;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.time.LocalDateTime;
-import java.util.List;
 
 @Slf4j
 @Service
